+++ lib/configParser.js
@@ -2,7 +2,7 @@ var path = require('path'),
     glob = require('glob'),
     util = require('util'),
     _ = require('lodash'),
-    protractor = require('./protractor.js');
+    helper = require('./util');
 
 // Coffee is required here to enable config files written in coffee-script.
 try {
@@ -32,7 +32,7 @@ var ConfigParser = function() {
       isVerbose: false,
       showColors: true,
       includeStackTrace: true,
-      stackFilter: protractor.filterStackTrace,
+      stackFilter: helper.filterStackTrace,
       defaultTimeoutInterval: (30 * 1000)
     },
     seleniumArgs: [],
