+++ lib/protractor.js
@@ -1,6 +1,7 @@
 var url = require('url');
 var util = require('util');
 var webdriver = require('selenium-webdriver');
+var helper = require('./util');
 
 var clientSideScripts = require('./clientsidescripts.js');
 var ProtractorBy = require('./locators.js').ProtractorBy;
@@ -13,17 +14,6 @@ var WEB_ELEMENT_FUNCTIONS = [
   'click', 'sendKeys', 'getTagName', 'getCssValue', 'getAttribute', 'getText',
   'getSize', 'getLocation', 'isEnabled', 'isSelected', 'submit', 'clear',
   'isDisplayed', 'getOuterHtml', 'getInnerHtml', 'getId'];
-
-var STACK_SUBSTRINGS_TO_FILTER = [
-  'node_modules/minijasminenode/lib/',
-  'node_modules/selenium-webdriver',
-  'at Module.',
-  'at Object.Module.',
-  'at Function.Module',
-  '(timers.js:',
-  'jasminewd/index.js',
-  'protractor/lib/'
-];
 
 var DEFAULT_RESET_URL = 'data:text/html,<html></html>';
 var DEFAULT_GET_PAGE_TIMEOUT = 10000;
@@ -1478,21 +1468,11 @@ Protractor.prototype.pause = function(opt_debugPort) {
 
     var getDescriptions = function(frameOrTask, descriptions) {
       if (frameOrTask.getDescription) {
-        var getRelevantStack = function(stack) {
-          return stack.filter(function(line) {
-            var include = true;
-            for (var i = 0; i < STACK_SUBSTRINGS_TO_FILTER.length; ++i) {
-              if (line.toString().indexOf(STACK_SUBSTRINGS_TO_FILTER[i]) !==
-                  -1) {
-                include = false;
-              }
-            }
-            return include;
-          });
-        };
+        var stacktrace = frameOrTask.snapshot_.getStacktrace();
+        stacktrace = stacktrace ? stacktrace.join('\n').trim() : ''; 
         descriptions.push({
           description: frameOrTask.getDescription(),
-          stack: getRelevantStack(frameOrTask.snapshot_.getStacktrace())
+          stack: helper.filterStackTrace(stacktrace)
         });
       } else {
         for (var i = 0; i < frameOrTask.children_.length; ++i) {
@@ -1510,8 +1490,8 @@ Protractor.prototype.pause = function(opt_debugPort) {
     var asString = '-- WebDriver control flow schedule \n';
     for (var i = 0; i < descriptions.length; ++i) {
       asString += ' |- ' + descriptions[i].description;
-      if (descriptions[i].stack.length) {
-        asString += '\n |---' + descriptions[i].stack.join('\n |---');
+      if (descriptions[i].stack) {
+        asString += '\n |---' + descriptions[i].stack.replace(/\n/g, '\n |---');
       }
       if (i != (descriptions.length - 1)) {
         asString += '\n';
@@ -1570,29 +1550,4 @@ exports.setInstance = function(ptor) {
  */
 exports.getInstance = function() {
   return instance;
-};
-
-/**
- * Utility function that filters a stack trace to be more readable. It removes
- * Jasmine test frames and webdriver promise resolution.
- * @param {string} text Original stack trace.
- * @return {string}
- */
-exports.filterStackTrace = function(text) {
-  if (!text) {
-    return text;
-  }
-  var lines = [];
-  text.split(/\n/).forEach(function(line) {
-    var include = true;
-    for (var i = 0; i < STACK_SUBSTRINGS_TO_FILTER.length; ++i) {
-      if (line.indexOf(STACK_SUBSTRINGS_TO_FILTER[i]) !== -1) {
-        include = false;
-      }
-    }
-    if (include) {
-      lines.push(line);
-    }
-  });
-  return lines.join('\n');
 };
